feat(harness): story-id-matcher + refactor metrics consumers — green [story-h6]
